change in the data type and reader query

+++ src/main/java/com/pgrg/springbatch/entity/Bonus.java
@@ -16,5 +16,5 @@ public class Bonus{
 
     private String choiceCategoryCd;
 
-    private BigDecimal pointsEarned;
+    private Long pointsEarned;
 }
